fix: 거리 및 소요 시간 필드 long 타입으로 변경

+++ src/main/java/nextstep/subway/domain/strategy/Dijkstra.java
@@ -44,10 +44,10 @@ public class Dijkstra implements ShortestPathStrategy {
     }
 
     @Override
-    public int findShortestValue(Station source, Station target) {
+    public long findShortestValue(Station source, Station target) {
         GraphPath shortestPath = getPath(source, target);
         validateExistPath(shortestPath);
-        return (int) shortestPath.getWeight();
+        return (long) shortestPath.getWeight();
     }
 
     private GraphPath getPath(Station source, Station target) {
