+++ src/main/java/nextstep/subway/domain/strategy/ShortestPathStrategy.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface ShortestPathStrategy {
     List<Station> findShortestPath(Station source, Station target);
-    int findShortestValue(Station source, Station target);
+    long findShortestValue(Station source, Station target);
 }
